refactor(#343): app config service 사용 및 s3 service import route 변경
