+++ src/boards/services/BoardImage.service.ts
@@ -3,7 +3,7 @@ import { CreateHelpMeBoardImageDto } from '@src/boards/dto/helpMeBoard/create.bo
 import { CreateMentorBoardImageDto } from '@src/boards/dto/mentorBoard/create.mentor.board.image.dto';
 import { BoardImageRepository } from '@src/boards/repository/boardImage.repository';
 import { MentorBoardService } from '@src/boards/services/mentor.board.service';
-import { S3Service } from '@src/common/s3/s3.service';
+import { S3Service } from '@src/common/s3/services/s3.service';
 import { HelpMeBoardImage } from '@src/entities/HelpMeBoardImage';
 
 @Injectable()
